requirements.txt
fastapi
uvicorn[standard]
motor
pymongo[srv]
python-dotenv